Split out the primitive definitions used by multiple evalautors into a `Generic` module.  Refactor the `EvalPrims` class away, pusing the `iteValue` operation into `BitWord` and demoting the `evalPrim` operation into an ordinary (non-typeclass) operation.
