+++ src/Cryptol/Eval/Reference.lhs
@@ -29,10 +29,10 @@
 > import Cryptol.ModuleSystem.Name (asPrim)
 > import Cryptol.TypeCheck.Solver.InfNat (Nat'(..), nAdd, nMin, nMul)
 > import Cryptol.TypeCheck.AST
+> import Cryptol.Eval.Generic (lg2)
 > import Cryptol.Eval.Monad (EvalError(..), PPOpts(..))
 > import Cryptol.Eval.Type (TValue(..), isTBit, evalValType, evalNumType, tvSeq)
 > import Cryptol.Eval.Value (mkBv, ppBV)
-> import Cryptol.Eval.Concrete (lg2)
 > import Cryptol.Utils.Ident (Ident, mkIdent)
 > import Cryptol.Utils.Panic (panic)
 > import Cryptol.Utils.PP
