+++ src/Cryptol/ModuleSystem/Base.hs
@@ -9,8 +9,9 @@
 -- This is the main driver---it provides entry points for the
 -- various passes.
 
-{-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE ImplicitParams #-}
+{-# LANGUAGE RecordWildCards #-}
 
 module Cryptol.ModuleSystem.Base where
 
@@ -25,7 +26,7 @@ import Cryptol.ModuleSystem.Env (lookupModule
                                 , ModulePath(..), modulePathLabel)
 import qualified Cryptol.Eval                 as E
 import qualified Cryptol.Eval.Value           as E
-import           Cryptol.Eval.Concrete ()
+import qualified Cryptol.Eval.Concrete as Concrete
 import qualified Cryptol.ModuleSystem.NamingEnv as R
 import qualified Cryptol.ModuleSystem.Renamer as R
 import qualified Cryptol.Parser               as P
@@ -199,6 +200,7 @@ doLoadModule isrc path fp pm0 =
      tcm <- optionalInstantiate =<< checkModule isrc path pm
 
      -- extend the eval env, unless a functor.
+     let ?evalPrim = Concrete.evalPrim
      unless (T.isParametrizedModule tcm) $ modifyEvalEnv (E.moduleEnv () tcm)
      loadedModule path fp tcm
 
@@ -543,6 +545,7 @@ evalExpr e = do
   env <- getEvalEnv
   denv <- getDynEnv
   evopts <- getEvalOpts
+  let ?evalPrim = Concrete.evalPrim
   io $ E.runEval evopts $ (E.evalExpr () (env <> deEnv denv) e)
 
 evalDecls :: [T.DeclGroup] -> ModuleM ()
@@ -551,6 +554,7 @@ evalDecls dgs = do
   denv <- getDynEnv
   evOpts <- getEvalOpts
   let env' = env <> deEnv denv
+  let ?evalPrim = Concrete.evalPrim
   deEnv' <- io $ E.runEval evOpts $ E.evalDecls () dgs env'
   let denv' = denv { deDecls = deDecls denv ++ dgs
                    , deEnv = deEnv'
